cast

+++ firma.h
@@ -7,7 +7,7 @@ uint64_t gewinn (const int64_t *const d, const size_t n) {
   current_max = 0;
   max_ending_here = 0;
   
-  for(i = 0; i < n; i++) {
+  for(i = 0; i < (int64_t)n; i++) {
     max_ending_here += *ptr;
     if (max_ending_here < 0) {
       max_ending_here = 0;
